refactor(non-lucide-icons)[OK-325] Move non-lucif-icons directory to common
