+++ client/src/app/user-boards/(board-management)/[boardId]/page.tsx
@@ -2,7 +2,6 @@ import React from "react";
 import BoardDetailsInfo from "@/app/user-boards/(board-management)/[boardId]/_components/BoardDetails";
 import { getBoardDetailsData } from "@/services/board/fetchBoardDetails";
 import { BoardDetailsResponse } from "@/interfaces/responses/board-details-response";
-import BoardDetailsError from "@/components/error/BoardDetailsError";
 
 interface BoardDetailsPageProps {
   params: {
